Ex4/requirements.txt
functions_framework==3.7.0
protobuf>=3.19.0,<5.0.0dev
google-cloud-storage
cloudevents
google-cloud-logging